% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMap1.R
\name{cmap1_score}
\alias{cmap1_score}
\title{CMAP1 connectivity score}
\usage{
cmap1_score(
  S_up,
  S_down,
  gene_names,
  pert_names,
  mat_R = NULL,
  cell_line = NULL,
  drg_type = NULL,
  list_ranked_R = NULL,
  return_raw = FALSE,
  ncpus = 1,
  touchstone_signatures = NULL,
  save_raw = FALSE,
  path = NULL,
  normalize = FALSE,
  permuted_pval = FALSE,
  permute_nb = 10000,
  padj_method = "BH"
)
}
\arguments{
\item{S_up}{vector of strings representing the subset of
up regulated query (ex. disease) genes defined based on a
user-specified threshold of fold-change and/or significance.}

\item{S_down}{vector of strings representing the subset of
down regulated query (ex. disease) genes defined based on a
user-specified threshold of fold-change and/or significance.}

\item{gene_names}{names of all the genes (rows of
the matrix \code{mat_R} or names of the elements in the list \code{list_ranked_R}).
Default NULL.}

\item{pert_names}{names of all the perturbagenes (colums of
the matrix \code{mat_R} or names of the elements in the list \code{list_ranked_R})}

\item{mat_R}{expression matrix or path from which the expression
matrix could be loaded. The matrix must have both colnames and
rownames. \link{test the loading from path}}

\item{cell_line}{character vector of the cell line in which the
signature (ex. molecule) from the DB was generated. (Default NULL)}

\item{drg_type}{character vector of the drug type in which the
signature (ex. molecule) from the DB was generated. (Default NULL)}

\item{list_ranked_R}{list of rank vectors or path from
which the list of rank vectors could be loaded.
The vectors should be named (gene names) so as the list (perturbagene names)
\link{test the loading from path}}

\item{return_raw}{boolean set to \code{TRUE} to return only the raw
connectivity scores. (Default FALSE)}

\item{ncpus}{number of cores used in the parallel calculations (Default 1).}

\item{touchstone_signatures}{names of the DB signatures (ex. perturbagenes)
considered as touchstones.}

\item{save_raw}{boolean set to \code{TRUE} to save the raw connectivity score
vector as an .RDS in a specific \code{path}. (Default FALSE)}

\item{path}{path used to save the raw connectivity scores if
save_raw is set to TRUE. (Default NULL)}

\item{normalize}{boolean set to \code{TRUE} to calculate the
normalized connectivity scores. (Default FALSE)}

\item{permuted_pval}{boolean set to \code{TRUE} to calculate the permuted pvalue
of the scores. (Default FALSE)}

\item{permute_nb}{number of permutation to calcule the pvalue
(Default 10000).}

\item{padj_method}{Method of adjustment of the pvalue. See stats::p.adjust
\code{method} parameter. (Default 'BH')}
}
\description{
Calculation of the raw or normalized connectivity score for
all the pertubation DB
}
