% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMap1.R
\name{cs}
\alias{cs}
\title{CMAP1 raw connectivity score}
\usage{
cs(S_up, S_down, R)
}
\arguments{
\item{S_up}{vector of strings representing the subset of
up regulated query (ex. disease) genes defined based on a
user-specified threshold of fold-change and/or significance.}

\item{S_down}{vector of strings representing the subset of
down regulated query (ex. disease) genes defined based on a
user-specified threshold of fold-change and/or significance.}

\item{R}{named vector with the drug perturbation ranks from the highest
to the lowest gene scores.}
}
\value{
raw connectivity score
}
\description{
CMAP1 raw connectivity score
}
