% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSS.R
\name{css_ordered}
\alias{css_ordered}
\title{CSS score for ordered disease query}
\usage{
css_ordered(R, sr_S)
}
\arguments{
\item{R}{named vector with drug perturbation data representing the drug gene
list of interest with expression information.}

\item{S}{named vector with disease expression data}
}
\value{
CSS score
(https://bmcbioinformatics.biomedcentral.com/articles/10.1186/1471-2105-9-258)
}
\description{
CSS score for ordered disease query
}
