% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSS.R
\name{css_score}
\alias{css_score}
\title{CSS score}
\usage{
css_score(
  pert_names,
  S = NULL,
  S_up = NULL,
  S_down = NULL,
  gene_names = NULL,
  mat_R = NULL,
  list_ranked_R = NULL,
  ncpus = 1,
  permuted_pval = FALSE,
  permute_nb = 10000,
  padj_method = "BH",
  unordered_S = FALSE
)
}
\arguments{
\item{pert_names}{names of all the perturbagenes (colums of
the matrix \code{mat_R} or names of the elements in the list \code{list_ranked_R})}

\item{S}{named vector with disease expression data.}

\item{S_up}{vector of strings representing the subset of
up regulated query (ex. disease) genes defined based on a
user-specified threshold of fold-change and/or significance.}

\item{S_down}{vector of strings representing the subset of
down regulated query (ex. disease) genes defined based on a
user-specified threshold of fold-change and/or significance.}

\item{gene_names}{names of all the genes (rows of
the matrix \code{mat_R} or names of the elements in the list \code{list_ranked_R}).
Default NULL.}

\item{mat_R}{expression matrix or path from which the expression
matrix could be loaded. The matrix must have both colnames and
rownames. \link{test the loading from path}}

\item{list_ranked_R}{list of rank vectors or path from
which the extreme list of rank vectors could be loaded.
The vectors should be named (gene names) so as the list (perturbagene names)
\link{test the loading from path}}

\item{ncpus}{number of cores used in the parallel calculations (Default 1).}

\item{permuted_pval}{boolean set to \code{TRUE} to calculate the permuted pvalue
of the scores. (Default FALSE)}

\item{permute_nb}{number of permutation to calcule the pvalue
(Default 10000).}

\item{padj_method}{Method of adjustment of the pvalue. See stats::p.adjust
\code{method} parameter. (Default 'BH')}

\item{unordered_S}{boolean set to TRUE if S is unordered and just contain
gene names}
}
\description{
Calculation of the CSS for all the pertubation DB
}
