% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMap1.R
\name{ES_cmap1}
\alias{ES_cmap1}
\title{Enrichment score using the signed one-sample KS test}
\usage{
ES_cmap1(Sx, R)
}
\arguments{
\item{Sx}{vector of strings representing the subset of
query (ex. disease) genes (S) with the most extreme gene scores
(either from the top or bottom) defined based on a
user-specified threshold of fold-change and/or significance.}

\item{R}{named vector with the drug perturbation ranks from the highest
to the lowest gene scores.}
}
\value{
enrichment score
}
\description{
CMap 1.0 uses a signed one-sample KS test to compare the
empirical distribution of the positions of SX genes in R
compared to a reference uniform distribution (of disease
genes in the drug gene list)
}
