% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMap2.R
\name{ES_gsea}
\alias{ES_gsea}
\title{Enrichment score based on GSEA}
\usage{
ES_gsea(Sx, R, wES)
}
\arguments{
\item{Sx}{vector of strings representing the subset of
query (ex. disease) genes (S) with the most extreme gene scores
(either from the top or bottom) defined based on a
user-specified threshold of fold-change and/or significance.}

\item{R}{named vector with the drug perturbation expression data
rank-ordered from the highest
to the lowest gene scores.}

\item{wES}{int weight assigned to the positions in R when
calculating the ES (default 1).}
}
\value{
enrichment score
}
\description{
The disease-drug ES in CMap 2.0 is based directly on GSEA’s
weighted signed two-sample KS statistic that compares the positions
of Sx genes to those of R − SX genes with the weight wES set to 1.
}
