% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixToRankedList.R
\name{matrixToRankedList}
\alias{matrixToRankedList}
\title{Transform an expression data matrix to a ranked list}
\usage{
matrixToRankedList(
  mat_R,
  ncpus = 1,
  expression_val = FALSE,
  signed_rank = FALSE
)
}
\arguments{
\item{mat_R}{expression data matrix with colnames (perturbagene)
and colnames (genes).}

\item{ncpus}{number of cores used in the parallel calculations (Default 1).}

\item{expression_val}{boolean set to \code{TRUE} to return the drug
perturbation expression data
rank-ordered from the highest to the lowest gene scores (CMAP2).}

\item{signed_rank}{boolean set to \code{TRUE} to return the drug perturbation
signed rank from the highest to the lowest absolute gene scores. (CSS)}
}
\description{
drug perturbation ranks from the highest
to the lowest gene scores.
}
