% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMap1.R
\name{normalization_cmap1}
\alias{normalization_cmap1}
\title{Normalized connectivity score for CMAP1. This normalization is
needed for CMAP1.}
\usage{
normalization_cmap1(cs, touchstone, cell_line = NULL, drg_type = NULL)
}
\arguments{
\item{cs}{vector of raw connectivity scores}

\item{touchstone}{vector of raw connectivity scores associated
to the touchstone signatures.}

\item{cell_line}{character vector of the cell line in which the
signature (ex. molecule) from the DB was generated. (Default NULL)}

\item{drg_type}{character vector of the drug type in which the
signature (ex. molecule) from the DB was generated. (Default NULL)}
}
\value{
normalized connectivity score
}
\description{
The final connectivity score is calculated by normalizing
the raw score by dividing by the maximum or minimum of raw
scores across treatment instances, depending on the sign of cs,
bringing it back to range between −1 and+ 1.
The use of \code{touchstone}, \code{cell_line} and \code{drg_type} parameters were
inspired from CMAP2 score, not necessary used in the original CMAP1 method.
}
