% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMap2.R
\name{normalization_cmap2}
\alias{normalization_cmap2}
\title{Normalized connectivity score for CMAP2}
\usage{
normalization_cmap2(wcs, touchstone, cell_line = NULL, drg_type = NULL)
}
\arguments{
\item{wcs}{vector of weighted connectivity scores}

\item{touchstone}{vector of weighted connectivity scores associated
to the touchstone signatures.}

\item{cell_line}{character vector of the cell line in which the
signature (ex. molecule) from the DB was generated. (Default NULL)}

\item{drg_type}{character vector of the drug type in which the
signature (ex. molecule) from the DB was generated. (Default NULL)}
}
\value{
normalized connectivity score
}
\description{
The NCS was developed to enable the comparison of WCS across
cell lines and drug type. Given the WCS for a disease in
relation to a specific drug of a type dt, tested in cell line c.
}
