% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMap2.R
\name{wcs}
\alias{wcs}
\title{CMAP2 weighted connectivity score}
\usage{
wcs(S_up, S_down, R, wES)
}
\arguments{
\item{S_up}{vector of strings representing the subset of
up regulated query (ex. disease) genes defined based on a
user-specified threshold of fold-change and/or significance.}

\item{S_down}{vector of strings representing the subset of
down regulated query (ex. disease) genes defined based on a
user-specified threshold of fold-change and/or significance.}

\item{R}{named vector with the drug perturbation expression data
rank-ordered from the highest to the lowest gene scores.}

\item{wES}{int weight assigned to the positions in R when
calculating the ES (default 1).}
}
\value{
weighted connectivity score
}
\description{
WCS ranges from -1 to +1, it is calculated using the enrichment scores for
the up and down regulated query lists.
}
