% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eXtremeScores.R
\name{XCor}
\alias{XCor}
\title{Pearson and eXtreme Pearson correlation}
\usage{
XCor(S, R, method = "pearson")
}
\arguments{
\item{S}{named vector with disease expression data}

\item{R}{named vector with drug perturbation expression data}

\item{method}{character to specify the use of \code{pearson} or \verb{spearman`` correlation. (Default }pearson`).}
}
\value{
XCor score (Pearson or Spearman correlation)
}
\description{
Pearson and eXtreme Pearson correlation
}
