% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eXtremeScores.R
\name{XCos}
\alias{XCos}
\title{Cosine and eXtreme Cosine correlation}
\usage{
XCos(S, R)
}
\arguments{
\item{S}{named vector with disease expression data}

\item{R}{named vector with drug perturbation expression data}
}
\value{
XCos score (cosine similarity)
}
\description{
Cosine and eXtreme Cosine correlation
}
