% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eXtremeScores.R
\name{XCos_score}
\alias{XCos_score}
\title{XCos score}
\usage{
XCos_score(
  S,
  pert_names,
  gene_names = NULL,
  mat_R = NULL,
  list_ranked_R = NULL,
  topN = NULL,
  ncpus = 1,
  permuted_pval = FALSE,
  permute_nb = 10000,
  padj_method = "BH"
)
}
\arguments{
\item{S}{named vector with disease expression data}

\item{pert_names}{names of all the perturbagenes (colums of
the matrix \code{mat_R} or names of the elements in the list \code{list_ranked_R})}

\item{gene_names}{names of all the genes (rows of
the matrix \code{mat_R} or names of the elements in the list \code{list_ranked_R}).
Default NULL.}

\item{mat_R}{expression matrix or path from which the expression
matrix could be loaded. The matrix must have both colnames and
rownames. \link{test the loading from path}}

\item{list_ranked_R}{list of rank vectors or path from
which the extreme list of rank vectors could be loaded.
The vectors should be named (gene names) so as the list (perturbagene names)
\link{test the loading from path}}

\item{topN}{integer to select the N extreme drug
perturbation expression data
rank-ordered from the highest to the lowest gene scores (eXtreme scores)}

\item{ncpus}{number of cores used in the parallel calculations (Default 1).}

\item{permuted_pval}{boolean set to \code{TRUE} to calculate the permuted pvalue
of the scores. (Default FALSE)}

\item{permute_nb}{number of permutation to calcule the pvalue
(Default 10000).}

\item{padj_method}{Method of adjustment of the pvalue. See stats::p.adjust
\code{method} parameter. (Default 'BH')}
}
\description{
Calculation of the XCos score for all the pertubation DB
}
