% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eXtremeScores.R
\name{XSum}
\alias{XSum}
\title{XSum extreme sum score}
\usage{
XSum(S_up, S_down, R)
}
\arguments{
\item{S_up}{vector of strings representing the subset of
up regulated query (ex. disease) genes defined based on a
user-specified threshold of fold-change and/or significance.}

\item{S_down}{vector of strings representing the subset of
down regulated query (ex. disease) genes defined based on a
user-specified threshold of fold-change and/or significance.}

\item{R}{named vector with the extreme drug perturbation
expression data rank-ordered from the highest to the lowest gene scores.}
}
\value{
Xsum
}
\description{
XSum extreme sum score
}
